Changed the TimeStamp type to long. refs #147

+++ WebServer.Tests/Controllers/PlaneDataControllerTest.cs
@@ -32,7 +32,7 @@ namespace WebServer.Tests.Controllers
         public void Test_SetMotor()
         {
             //Act
-            var result = SystemUnderTest.SetMotor(new Dictionary<int, int> {{1, 1}});
+            var result = SystemUnderTest.SetMotor(new Dictionary<long, int> {{1, 1}});
 
             //Assert
             Assert.That(()=>_smartPlaneTestUser.MotorDatas.First().Value,Is.EqualTo(1));
@@ -44,7 +44,7 @@ namespace WebServer.Tests.Controllers
         public void Test_SetMotor_NoData_Empty_List()
         {
             //Act
-            var result = SystemUnderTest.SetMotor(new Dictionary<int, int>());
+            var result = SystemUnderTest.SetMotor(new Dictionary<long, int>());
 
             //Assert    
             Assert.That(() => result.StatusCode, Is.EqualTo(HttpStatusCode.NoContent));
@@ -64,7 +64,7 @@ namespace WebServer.Tests.Controllers
         public void Test_SetRudder()
         {
             //Act
-            var result = SystemUnderTest.SetRudder(new Dictionary<int, int> { { 1, 1 } });
+            var result = SystemUnderTest.SetRudder(new Dictionary<long, int> { { 1, 1 } });
             
             //Assert
             Assert.That(()=>_smartPlaneTestUser.RudderDatas.First().Value,Is.EqualTo(1));
@@ -76,7 +76,7 @@ namespace WebServer.Tests.Controllers
         public void Test_SetRudder_NoData_Empty_List()
         {
             //Act
-            var result = SystemUnderTest.SetRudder(new Dictionary<int, int>());
+            var result = SystemUnderTest.SetRudder(new Dictionary<long, int>());
 
             //Assert
             Assert.That(() => result.StatusCode, Is.EqualTo(HttpStatusCode.NoContent));
@@ -96,7 +96,7 @@ namespace WebServer.Tests.Controllers
         public void Test_SetIsConnected()
         {
             //Act
-            var result = SystemUnderTest.SetIsConnected(new Dictionary<int, bool> { { 1, true } });
+            var result = SystemUnderTest.SetIsConnected(new Dictionary<long, bool> { { 1, true } });
 
             //Assert
             Assert.That(()=>_smartPlaneTestUser.ConnectedDatas.First().IsConnected,Is.EqualTo(true));
@@ -108,7 +108,7 @@ namespace WebServer.Tests.Controllers
         public void Test_SetIsConnected_NoData_Empty_List()
         {
             //Act
-            var result = SystemUnderTest.SetIsConnected(new Dictionary<int, bool>());
+            var result = SystemUnderTest.SetIsConnected(new Dictionary<long, bool>());
 
             //Assert
             Assert.That(() => result.StatusCode, Is.EqualTo(HttpStatusCode.NoContent));
